% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanss3.R
\name{clusterAssignment}
\alias{clusterAssignment}
\title{Assignment of clusters based on distance}
\usage{
clusterAssignment(x, clusterMatrix, currentCenters, centersBefore)
}
\arguments{
\item{x}{Numeric data matrix}

\item{clusterMatrix}{Matrix for clustering purposes}

\item{currentCenters}{Current mean values of each cluster}

\item{centersBefore}{Centers for current iteration-1}
}
\value{
Returns a matrix with the assignment of n observations to k clusters.
}
\description{
Points are assigned to the nearest cluster based on the squared
Euclidean distance.
}
