% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanss3.R
\name{initialCenters}
\alias{initialCenters}
\title{Creation of initial centers}
\usage{
initialCenters(x, k)
}
\arguments{
\item{x}{Numeric data matrix}

\item{k}{Either a number or matrix to define initial centers}
}
\value{
Returns the initial k centers.
}
\description{
Initial centers are created based on input or random selection.
}
