% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanss3.R
\name{myMeans}
\alias{myMeans}
\title{K-Means Clustering}
\usage{
myMeans(x, k, maxIteration = 10, stopCrit = 0)
}
\arguments{
\item{x}{x is a numeric data matrix, or an object that can be turned into a
matrix, i.e. vector or dataframe. These may only contain numeric values.}

\item{k}{k is either a number or matrix. If k is a number, a random set of
rows in x is chosen as the initial centers. If k is a matrix a set of
initial cluster centres is specified by the user.}

\item{maxIteration}{The maximum number of allowed iterations before the
algorithm stops.}

\item{stopCrit}{Threshold for mean difference of centers that stops the
algorithm}
}
\value{
The algorithm returns k clusters for the specified data x.
}
\description{
This package allows users to perform k-means clustering on a
data matrix using Lloyd's (1982) algorithm.
}
\examples{
x <- cbind((sample(1:100, 20)), sample(1:100, 20))
# create four clusters for data x with random initial centers
kClustering <- myMeans(x, 4)

# choose initial centers manually
kInitial <- matrix(c(20, 35, 40, 65), nrow = 2)
# create two clusters for x with pre-defined centers
kClustering <- myMeans(x, kInitial)
}
