% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethode.R
\name{plot.myMeans}
\alias{plot.myMeans}
\title{Plot Method for S3 Class \code{myMeans}}
\usage{
\method{plot}{myMeans}(x, ...)
}
\arguments{
\item{x}{An object of class myMeans \code{myMeans}}

\item{...}{Additional arguments}
}
\value{
A two-dimensional plot is returned.
}
\description{
plot() can be used to visualize the clustered data. Depending
on the dimensions of the input data, the user may be required to select
which columns are to be graphed.
}
