% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethode.R
\name{print.myMeans}
\alias{print.myMeans}
\title{Print Method for S3 Class \code{myMeans}}
\usage{
\method{print}{myMeans}(x, ...)
}
\arguments{
\item{x}{An object of class \code{myMeans}}

\item{...}{Additional arguments}
}
\value{
A text consisting of different components is returned.
}
\description{
Function to print k-means output less verbose
}
