% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanss3.R
\name{sanityCheck}
\alias{sanityCheck}
\title{Checking for sanity}
\usage{
sanityCheck(x, k, maxIteration, stopCrit)
}
\arguments{
\item{x}{Numeric data matrix}

\item{k}{Either a number or matrix to define initial centers}

\item{maxIteration}{The maximum allowed number of iterations}

\item{stopCrit}{Stopping criterion of algorithm}
}
\value{
Stops the execution of the myMeans function if inputs do not satisfy
the conditions.
}
\description{
The inputs are checked for validity and correctness.
}
