% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethode.R
\name{summary.myMeans}
\alias{summary.myMeans}
\title{Summary Method for S3 Class \code{myMeans}}
\usage{
\method{summary}{myMeans}(object, ...)
}
\arguments{
\item{object}{An object of class \code{myMeans}}

\item{...}{Additional arguments}
}
\value{
Returns a summary of the input object.
}
\description{
Function to return a summary of components.
}
