% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanss3.R
\name{totalSumOfSquares}
\alias{totalSumOfSquares}
\title{Calculation of total sum of squares}
\usage{
totalSumOfSquares(x)
}
\arguments{
\item{x}{data matrix}
}
\value{
returns the total sum of squares metric.
}
\description{
This function calculates the total sum of squares.
}
