% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanss3.R
\name{withinSumOfSquares}
\alias{withinSumOfSquares}
\title{Calculation of within-cluster sum of squares}
\usage{
withinSumOfSquares(x, currentCenters, clusters)
}
\arguments{
\item{x}{Numeric data matrix}

\item{currentCenters}{Mean values of each center}

\item{clusters}{Cluster assignment of each data point}
}
\value{
Returns the within-cluster sum of square metric.
}
\description{
This function calculates the within-cluster sum of squares.
}
